#585: Removed fields from client and account. Added softdeleteable to client
