+++ config/packages/stof_doctrine_extensions.yaml
@@ -6,3 +6,4 @@ stof_doctrine_extensions:
         default:
             timestampable: true
             blameable: true
+            softdeleteable: true
