+++ templates/client/index.html.twig
@@ -69,7 +69,7 @@
             </td>
         </tr>
         <tr class="table-tr">
-            <td class="table-td" colspan="6" class="text-right">
+            <td class="table-td" colspan="{{ numberOfColumns }}" class="text-right">
                 {{ knp_pagination_render(clients) }}
             </td>
         </tr>
